requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
deepl==1.20.0
idna==3.10
keyboard==0.13.5
MouseInfo==0.1.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
pytweening==1.2.0
requests==2.32.3
urllib3==2.2.3
